chore: disable form
disable form buttons, backend submission handler

122
